fixed the rest of the compiler warnings. It was complaining about casting an NSInteger to an Int value, so I explicitly casted to (long) instead. This issue was appearing in the "time interval" calculations when converting the NSTimeInterval to readable hours and minutes for display in the UI on multiple screens.

+++ EditDayViewController.m
@@ -53,7 +53,7 @@
     NSInteger ti = (NSInteger)interval;
     NSInteger minutes = (ti / 60) % 60;
     NSInteger hour = (ti / 3600);
-    return [NSString stringWithFormat:@"%i:%02i", hour, minutes];
+    return [NSString stringWithFormat:@"%li:%02li", (long)hour, (long)minutes];
 }
 
 - (void)didReceiveMemoryWarning
